refactor: update event api
