+++ app/adapter/output/slack.py
